api/src/main/kotlin/tournament/api/controller/TournamentController.kt
package tournament.api.controller

import io.micronaut.http.annotation.Controller
import io.micronaut.http.annotation.Get
import io.micronaut.validation.Validated
import io.reactivex.Single
import tournament.api.repository.Tournament
import tournament.api.service.TournamentService
import javax.annotation.security.PermitAll

@Validated
@PermitAll
@Controller("/tournament")
class TournamentController(private val service: TournamentService) {

    @Get("/list")
    fun getTournaments(): Single<List<Tournament>> {
        return Single.just(service.getTournaments())
    }


}